# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TokenInsights REST API
  version: 2.0.0
  description:
    Query dataset-scoped usage and ingest sanitized collector evidence. Personal servers are
    unauthenticated; hosted servers authenticate bearer tokens or read-only browser sessions.
    Servers interpret evidence and never scan host artifacts.
...
paths:
...
  /api/v2/instance:
    get:
      security:
        - bearerAuth: []
        - browserSession: []
        - {}
      operationId: getInstanceV2
      summary: Describe this TokenInsights server
      tags:
        - Instance
      responses:
        '200':
          description: Server versions, saved data hostname, capabilities, and viewer defaults
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceResponseV2'
        default:
          $ref: '#/components/responses/Error'
...
  /api/v2/usage:
    get:
      security:
        - bearerAuth: []
        - browserSession: []
        - {}
      operationId: getUsageV2
      summary: Query canonical token usage
      description: 'Returns a sorted, paginated table, complete filtered summary, and chart

        from one validated read transaction. Date bounds use server-local dates.

        '
      tags:
        - Usage
      parameters:
        - $ref: '#/components/parameters/PeriodFilter'
        - $ref: '#/components/parameters/BucketFilter'
        - $ref: '#/components/parameters/FromFilter'
        - $ref: '#/components/parameters/ToFilter'
        - $ref: '#/components/parameters/ProviderFilter'
        - $ref: '#/components/parameters/ModelFilter'
        - $ref: '#/components/parameters/HarnessFilter'
        - $ref: '#/components/parameters/SessionFilter'
        - $ref: '#/components/parameters/RepositoryFilter'
        - $ref: '#/components/parameters/DirectoryFilter'
        - name: tab
          in: query
          description: Aggregation shown in table and chart rows.
          schema:
            allOf:
              - $ref: '#/components/schemas/UsageTab'
            default: tokens
        - name: locationGroup
          in: query
          description: Repo-tab grouping by repository or directory.
          schema:
            $ref: '#/components/schemas/LocationGroup'
        - name: sort
          in: query
          description: Sort field. Context-only fields are valid only for the context tab.
          schema:
            $ref: '#/components/schemas/SortField'
        - name: direction
          in: query
          schema:
            $ref: '#/components/schemas/SortDirection'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - $ref: '#/components/parameters/QualityFilter'
      responses:
        '200':
          description: Canonical usage result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponseV2'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Error'
  /api/v2/usage/facets:
    get:
      security:
        - bearerAuth: []
        - browserSession: []
        - {}
      operationId: getUsageFacetsV2
      summary: Query available usage filters
      description: 'Each facet applies the other active filters while omitting its own.

        Session search is a case-insensitive literal substring search capped at

        100 results.

        '
      tags:
        - Usage
      parameters:
        - $ref: '#/components/parameters/PeriodFilter'
        - $ref: '#/components/parameters/BucketFilter'
        - $ref: '#/components/parameters/FromFilter'
        - $ref: '#/components/parameters/ToFilter'
        - $ref: '#/components/parameters/ProviderFilter'
        - $ref: '#/components/parameters/ModelFilter'
        - $ref: '#/components/parameters/HarnessFilter'
        - $ref: '#/components/parameters/SessionFilter'
        - $ref: '#/components/parameters/RepositoryFilter'
        - $ref: '#/components/parameters/DirectoryFilter'
        - name: search
          in: query
          description: Literal substring used to filter session IDs.
          schema:
            type: string
            default: ''
        - name: tab
          in: query
          description: Active aggregation. Repository/directory filters require the repo tab.
          schema:
            allOf:
              - $ref: '#/components/schemas/UsageTab'
            default: tokens
        - $ref: '#/components/parameters/QualityFilter'
      responses:
        '200':
          description: Available dimension values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageFacetsResponseV2'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Error'
  /api/v3/ingestion/capabilities:
    get:
      security:
        - bearerAuth: []
        - {}
      operationId: getRawIngestionCapabilitiesV3
      summary: Read raw ingestion contract and durable database identity
      tags:
        - Ingestion
      responses:
        '200':
          description: Acceptance-only ingestion contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawCapabilitiesV3'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '401':
          description: Hosted authentication required; no acceptance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '403':
          description: Hosted ingestion permission required; browser sessions are read-only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '429':
          description: Admission limit reached; retry exact bytes after backoff.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
      description:
        Describes durable acceptance limits and the effective server-owned dataset. Hosted requests
        require an ingest-scoped bearer token. Personal managed ingestion uses private Unix transport.
  /api/v3/ingestion/batches:
    post:
      security:
        - bearerAuth: []
        - {}
      operationId: ingestRawBatchV3
      summary: Durably accept immutable sanitized source evidence
      description:
        Raw evidence, submitted item mappings, immutable receipt and pending work commit together.
        Duplicate evidence succeeds without double counting. Processing runs asynchronously. Database and
        authenticated dataset binding are verified before acceptance. Hosted ingestion requires an ingest-scoped
        bearer token. Managed personal HTTP is query-only; ingestion uses private Unix transport. Metadata
        strings are limited to 256 UTF-8 bytes.
      tags:
        - Ingestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RawBatchV3'
      responses:
        '200':
          description:
            Durably accepted; all items terminal for current processing generation and scope
            revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawAcceptanceV3'
        '202':
          description: Durably accepted; processing remains asynchronous.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawAcceptanceV3'
        '400':
          description: Invalid JSON or private/unknown fields; no acceptance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '409':
          description: Batch bytes, sequence binding or database identity conflict; no acceptance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '413':
          description: Body exceeds 1 MiB; no acceptance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '422':
          description: Unsupported contract; no acceptance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '503':
          description: Cannot durably accept; retry exact bytes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '401':
          description: Hosted authentication required; no acceptance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '403':
          description: Hosted ingestion permission required; browser sessions are read-only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '429':
          description: Admission limit reached; retry exact bytes after backoff.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
...
  /api/v1/instance:
    get:
      operationId: getInstance
      summary: Describe this TokenInsights server
      tags:
        - Instance
      responses:
        '200':
          description: Server versions, saved data hostname, capabilities, and viewer defaults
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceResponse'
              example:
                apiVersion: v1
                serverVersion: 1.0.0
                hostname: workstation
                timezone: Australia/Sydney
                capabilities:
                  - usage
                  - facets
                  - ingestion
                defaults:
                  period: month
                  bucket: day
                  from: ''
                  to: ''
                  providers: []
                  models: []
                  harnesses: []
                  sessions: []
        default:
          $ref: '#/components/responses/Error'
  /api/v1/sync:
    get:
      operationId: getSync
      summary: Read ingestion readiness and committed revision
      tags:
        - Sync
      responses:
        '200':
          description: Saved server state; ready, idle, and no collector progress.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResponse'
        default:
          $ref: '#/components/responses/Error'
      description: Read-only compatibility endpoint. It never starts collection. POST is unsupported.
  /api/v1/usage:
    get:
      operationId: getUsage
      summary: Query canonical token usage
      description: 'Returns a sorted, paginated table, complete filtered summary, and chart

        from one validated read transaction. Date bounds use server-local dates.

        '
      tags:
        - Usage
      parameters:
        - $ref: '#/components/parameters/PeriodFilter'
        - $ref: '#/components/parameters/BucketFilter'
        - $ref: '#/components/parameters/FromFilter'
        - $ref: '#/components/parameters/ToFilter'
        - $ref: '#/components/parameters/ProviderFilter'
        - $ref: '#/components/parameters/ModelFilter'
        - $ref: '#/components/parameters/HarnessFilter'
        - $ref: '#/components/parameters/SessionFilter'
        - $ref: '#/components/parameters/RepositoryFilter'
        - $ref: '#/components/parameters/DirectoryFilter'
        - name: tab
          in: query
          description: Aggregation shown in table and chart rows.
          schema:
            allOf:
              - $ref: '#/components/schemas/UsageTab'
            default: tokens
        - name: locationGroup
          in: query
          description: Repo-tab grouping by repository or directory.
          schema:
            $ref: '#/components/schemas/LocationGroup'
        - name: sort
          in: query
          description: Sort field. Context-only fields are valid only for the context tab.
          schema:
            $ref: '#/components/schemas/SortField'
        - name: direction
          in: query
          schema:
            $ref: '#/components/schemas/SortDirection'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - $ref: '#/components/parameters/QualityFilter'
      responses:
        '200':
          description: Canonical usage result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/usage/facets:
    get:
      operationId: getUsageFacets
      summary: Query available usage filters
      description: 'Each facet applies the other active filters while omitting its own.

        Session search is a case-insensitive literal substring search capped at

        100 results.

        '
      tags:
        - Usage
      parameters:
        - $ref: '#/components/parameters/PeriodFilter'
        - $ref: '#/components/parameters/BucketFilter'
        - $ref: '#/components/parameters/FromFilter'
        - $ref: '#/components/parameters/ToFilter'
        - $ref: '#/components/parameters/ProviderFilter'
        - $ref: '#/components/parameters/ModelFilter'
        - $ref: '#/components/parameters/HarnessFilter'
        - $ref: '#/components/parameters/SessionFilter'
        - $ref: '#/components/parameters/RepositoryFilter'
        - $ref: '#/components/parameters/DirectoryFilter'
        - name: search
          in: query
          description: Literal substring used to filter session IDs.
          schema:
            type: string
            default: ''
        - name: tab
          in: query
          description: Active aggregation. Repository/directory filters require the repo tab.
          schema:
            allOf:
              - $ref: '#/components/schemas/UsageTab'
            default: tokens
        - $ref: '#/components/parameters/QualityFilter'
      responses:
        '200':
          description: Available dimension values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageFacetsResponse'
              example:
                providers:
                  - openai
                models:
                  - gpt-5
                harnesses:
                  - codex
                sessions:
                  - session-123
                repositories: []
                directories: []
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/ingestion/capabilities:
    get:
      operationId: getIngestionCapabilities
      summary: Read normalized ingestion compatibility and database identity
      tags:
        - Ingestion
      responses:
        '200':
          description: Compatible protocol and current durable server database identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionCapabilities'
        '503':
          description: Server storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
      deprecated: true
  /api/v1/ingestion/batches:
    post:
      operationId: ingestBatch
      summary: Retain a legacy normalized publication batch
      description:
        Compatibility delivery for retained protocol-1 journal requests. Facts are stored as
        legacy baselines with synchronous receipts. New collectors use protocol 2. Local writes use private
        Unix transport. Matching proven raw contributions replace baselines by identity; unmatched baselines
        remain queryable.
      tags:
        - Ingestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicationBatch'
      responses:
        '200':
          description: Durable committed receipt, including exact replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionReceipt'
        '400':
          description: Invalid JSON, identity, counters, or normalized values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '401':
          description: Reserved for future authenticated deployments; current deployments are unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '409':
          description: Database binding, fact payload, reference, or batch identity conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '413':
          description: Request body exceeds 1 MiB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '503':
          description: Admission, writer, or storage unavailable; retry exact bytes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '500':
          description: Transaction failed with no partial acceptance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '422':
          description: Unsupported protocol, identity, or semantics version; no mutation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
      deprecated: true
  /api/v2/ingestion/capabilities:
    get:
      operationId: getRawIngestionCapabilities
      summary: Read raw ingestion contract and durable database identity
      tags:
        - Ingestion
      responses:
        '200':
          description: Acceptance-only ingestion contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawCapabilities'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
  /api/v2/ingestion/batches:
    post:
      operationId: ingestRawBatch
      summary: Durably accept immutable sanitized source evidence
      description:
        Raw evidence, every submitted item mapping, immutable receipt and pending work commit
        together. Duplicate evidence succeeds. Processing runs asynchronously. Local ingestion is private
        Unix transport; public local HTTP exposes queries only. Remote authentication remains a separate
        deployment concern. Metadata strings are limited to 256 UTF-8 bytes.
      tags:
        - Ingestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RawBatch'
      responses:
        '200':
          description:
            Durably accepted; all items terminal for current processing generation and scope
            revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawAcceptance'
        '202':
          description: Durably accepted; processing remains asynchronous.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawAcceptance'
        '400':
          description: Invalid JSON or private/unknown fields; no acceptance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '409':
          description: Batch bytes, sequence binding or database identity conflict; no acceptance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '413':
          description: Body exceeds 1 MiB; no acceptance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '422':
          description: Unsupported contract; no acceptance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '503':
          description: Cannot durably accept; retry exact bytes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Hosted token; effective dataset and read/ingest scopes resolved by
        the server.
    browserSession:
      type: apiKey
      in: cookie
      name: tokeninsights_session
      description: Read-only, HttpOnly browser session established by POST /api/v2/auth/session.
  parameters:
    PeriodFilter:
      name: period
      in: query
      schema:
        allOf:
          - $ref: '#/components/schemas/Period'
        default: month
    BucketFilter:
      name: bucket
      in: query
      schema:
        allOf:
          - $ref: '#/components/schemas/Bucket'
        default: day
    FromFilter:
      name: from
      in: query
      description: Inclusive lower local-date bound; replaces period when supplied.
      schema:
        type: string
        format: date
        x-go-type: string
...
  responses:
...
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    InstanceResponseV2:
      type: object
      additionalProperties: false
      required:
        - instanceId
        - dataEpoch
        - dataReadiness
        - apiVersion
        - serverVersion
        - hostname
        - timezone
        - capabilities
        - defaults
        - datasetId
        - serverKind
        - permissions
      properties:
        instanceId:
          type: string
          minLength: 1
        dataEpoch:
          type: string
          description: Durable database identity; nonempty when dataReadiness is ready.
        dataReadiness:
          type: string
          enum:
            - ready
            - metadata
            - recovery
            - rebuild
            - unavailable
        apiVersion:
          type: string
          enum:
            - v2
        serverVersion:
          type: string
          minLength: 1
        hostname:
          description:
            'Producer hostname: unknown before ingestion; multiple machines when distinct producer
            labels are present.'
          type: string
          minLength: 1
        timezone:
          type: string
          minLength: 1
          description:
            Server reporting timezone, preferably an IANA name with historical DST rules. Unnamed
            system zones use an explicit UTC offset fallback.
        capabilities:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
            description:
              'Unknown capabilities are ignored by clients. Known names: usage, facets, web-dashboard,
              raw-ingestion, terminal-dashboard, collector-progress, reprocess.'
        defaults:
          $ref: '#/components/schemas/Selection'
        datasetId:
          type: string
          minLength: 1
          maxLength: 256
        serverKind:
          type: string
          enum:
            - personal
            - hosted
        permissions:
          type: array
          uniqueItems: true
          items:
            type: string
            enum:
              - read
              - ingest
...